feat: 세션 유무에 따라 사이드바 프로필 영역 조건부 렌더링
- 사이드바 내부에서 서버 사이드로 세션을 확인해 재사용성을 확보하기 위해 Sidebar를 서버 컴포넌트로 전환
- usePathname을 사용하는 메뉴 이동 버튼(SidebarNav)과 드롭다운/로그아웃을 포함한 프로필 영역(SidebarProfile)을 각각 별도 클라이언트 컴포넌트로 분리
- 세션이 없으 SidebarProfile을 마운트하지 않도록 처리
